Solved Issue when using blur layer
When adding a blurEffect layer to the current UIView, SweetAlert pops the alert behind the blurLayer and you see alert blurred and not in focus (firstResponder).

To fix it, changes from this

let window = UIApplication.sharedApplication().keyWindow?.subviews.first as! UIView

to this

let window = UIApplication.sharedApplication().keyWindow?.viewForBaselineLayout() as UIView!